# doc/api.yaml
openapi: 3.0.3
info:
  title: WasaText API
  description: |-
    This OpenAPI document describes the simplified login API.
    Copy and paste the API from the `paths` key to your OpenAPI document.
  version: "A"
...
paths:
...
  /user/{user_id}/conversation/{dest_id}: 
    parameters:
      - $ref: "#/components/parameters/userid"
      - $ref: "#/components/parameters/destid"
    put:
      security:
        - bearerAuth: []  
      tags: [User]
      summary: create a new chat
      description: |-
        create a new chat
      operationId: createChat
      requestBody:
        content: 
          application/json:
            schema:
              type: object
              description: message
              properties:
                msgid: 
                  $ref: "#/components/schemas/Messages"

      responses:
        "201": 
          description: chat created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"  
        "500":
          $ref: "#/components/responses/InternalServerError"

    get:
      security:
        - bearerAuth: []
      tags: [User]
      operationId: getConversation
      description: get a specific conversation 
      summary: return a single conveersation
      responses:
        "200":
          description: Conversation retrieved by chat_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"  
        "500":
          $ref: "#/components/responses/InternalServerError" 
...
  /user/{user_id}/conversation/{chat_id}/messages/{message_id}/comments:
    parameters:
      - $ref: "#/components/parameters/userid"
      - $ref: "#/components/parameters/chatid"
      - $ref: "#/components/parameters/messageid"

    put:
      operationId: commentMessage
      description: Comment on a single message
      summary: comment a single message with reaction
      tags: [User]
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Reaction"
      responses:
        "201": 
          description: Replied successfully
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Reaction"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"  
        "500":
          $ref: "#/components/responses/InternalServerError" 
    delete:
      operationId: uncommentMessage
      description: Delete a comment of a message
      summary: delete a comment of a message
      tags: [User]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Comment deleted
          content: 
            text/plain:
              schema: 
                description: comment deleted
                type: string
                pattern: '^.*?$'
                minLength: 5
                maxLength: 100
                example: "comment deleted  correctly"

        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"  
        "500":
          $ref: "#/components/responses/InternalServerError"       
...
components:
  securitySchemes:
    bearerAuth: 
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    userid:
      description: userID
      name: user_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Identifier"
    chatid:
      description: chat id
      name: chat_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Identifier"
    messageid:
      description: messageid
      name: message_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Identifier"
...
    destid:
      description: identifier
      name: dest_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Identifier"
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
      content: 
        text/plain:
          schema: 
            description: Unauthorized error
            type: string
            pattern: '^.*?$'
            example: "Unathorized Error"
            minLength: 12
            maxLength: 30
    BadRequest:
      description: Bad request
      content: 
        text/plain:
          schema: 
            description: Bad request
            type: string
            pattern: '^.*?$'
            example: "bad request"
            minLength: 11
            maxLength: 30
    ForbiddenError:
      description: User is not allowed to perform action 
      content: 
        text/plain:
          schema: 
            description: forbidden error
            type: string
            pattern: '^.*?$'
            example: "Forbuidden Error"
            minLength: 9
            maxLength: 30 
    InternalServerError:
      description: Internal server error
      content: 
        text/plain:
          schema: 
            description: Internal server error
            type: string
            pattern: '^.*?$'
            example: "Internal server Error"
            minLength: 21
            maxLength: 30  
  schemas:
    User: 
      type: object
      description: user object
      properties:
        user_id:
          $ref: "#/components/schemas/Identifier"
        username:
          $ref: "#/components/schemas/Username"
        photouser:
            $ref: "#/components/schemas/Photo"
    Username:
      type: object
      description: object user
      
      properties:
        syntaxname:
          type: string
          description: username
          pattern: '^.*?$'
          example: "Mario"
          minLength: 3
          maxLength: 16
    Identifier:
      type: integer
      description: identifier
      minimum: 1
      maximum: 100000
      example: 110
    ConversationGroup:
      type: object
      description: group chat
      properties:
        group_id:
          $ref: "#/components/schemas/Identifier"
        
        chat_id:
          $ref: "#/components/schemas/Identifier"      
...
    ConversationPrivate:
      type: object
      description: private chat
      properties:
        chatid:
          $ref: "#/components/schemas/Identifier"
        user:
          $ref: "#/components/schemas/User" 
    Conversation:
      description: Schema of conversation
      oneOf:
        - $ref: "#/components/schemas/ConversationPrivate"
        - $ref: "#/components/schemas/ConversationGroup"
...
    Messages: 
      description: message schema
      anyOf:
        - $ref: "#/components/schemas/Txtmsg"
        - $ref: "#/components/schemas/Photomsg"
    Photomsg:
      type: object
      description: photo message
      properties:
        photoid: 
          $ref: "#/components/schemas/Identifier"
        
        photo:
          $ref: "#/components/schemas/Photo"
...
    Txtmsg:
      type: object
      description: text message
      properties:
        txt:
          type: string
          description: text 
          pattern: '^.*?$'
          example: "Ciao sono Mario!"
          minLength: 1
          maxLength: 100000
        date:
          type: string
          description: date time
          format: date-time
          example: "2020-12-09T10:45:03Z"
          minLength: 1
          maxLength: 100000
        id:
          $ref: "#/components/schemas/Identifier"
        sender:
          description: sender
          type: string
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
        receiver:
          type: string
          description: receiver
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
    Photo:
        type: string
        description: foto
        format: byte
        minLength: 1
        maxLength: 1000000000      
    Reaction:
      type: object
      description: reaction 
      properties:
        emoticon:
          type: string
          description: emoticon
          pattern: '^.*?$'
          minLength: 1
          maxLength: 2
        id:
          $ref : "#/components/schemas/Identifier"